Quant.Explorer.HttpClientConfig (quant v0.1.0-alpha.1)

Configuration module for selecting the appropriate HTTP client based on environment.

Functions

decode_json(body)

Decodes JSON using the configured client.

get(url, params \\ %{}, opts \\ [])

Makes an HTTP GET request using the configured client.

http_client()

Returns the HTTP client module to use based on the current environment. In test environment with mocking enabled, returns the test client. Otherwise returns the standard HTTP client.

post(url, body, opts \\ [])

Makes an HTTP POST request using the configured client.